requirements.txt
asyncio
paho.mqtt
astropy
